% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.data.R
\name{changeColumnNames}
\alias{changeColumnNames}
\title{Re-formats column names}
\usage{
changeColumnNames(col)
}
\arguments{
\item{col}{Vector of column names}
}
\value{
Returns a newly formatted vector of column names
}
\description{
Changes the way the column names are formatted from nxn.y to the n[x>y]n format
}
\keyword{internal}
