% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.to.sigs.input.R
\name{findComp}
\alias{findComp}
\title{Returns complement of a sequence}
\usage{
findComp(tri, nr = FALSE)
}
\arguments{
\item{tri}{Character vector of "A", "G", "C", or "T"}

\item{nr}{If TRUE returns only the complement of the input sequence}
}
\value{
Returns a character vector containing the complemented input sequence
}
\description{
Given an input sequence, returns either the reverse complement (default) or
complement of the sequnce (nr = TRUE)
}
\keyword{internal}

