% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.scripts.R
\name{findSeed}
\alias{findSeed}
\title{Seeds weight matrix}
\usage{
findSeed(tumor, signatures)
}
\arguments{
\item{tumor}{Actual trinucleotide context fractions}

\item{signatures}{Signatures matrix}
}
\value{
Returns the index corresponding to the signature that best describes
  the input data
}
\description{
Determines which single signature results in the lowest sum-squared error and
uses that as a seed to start from
}
\keyword{internal}
