% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{formatContexts}
\alias{formatContexts}
\title{Formats the data for use in plotSignatures()}
\usage{
formatContexts(contexts)
}
\arguments{
\item{contexts}{One of the entries in the output list from whichSignatures()}
}
\value{
Returns a data frame with sample.id, full_context, fraction, and 
  mutation as column names
}
\description{
Makes the contexts useable in plotSignatures()
}
\keyword{internal}
