% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.scripts.R
\name{getError}
\alias{getError}
\title{Calculates error}
\usage{
getError(tumor, signatures, w)
}
\arguments{
\item{tumor}{Actual trinucleotide context fractions}

\item{signatures}{Signatures matrix}

\item{w}{Weights matrix}
}
\value{
Returns the sum squared error between calculated and actual
  trinucleotide context fractions
}
\description{
Finds error between calculated trinucleotide context fractions and inputted
ones
}
\keyword{internal}
