% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.data.R
\name{getTriContextFraction}
\alias{getTriContextFraction}
\title{Calculates trinucleotide context fraction}
\usage{
getTriContextFraction(
  mut.counts,
  trimer.counts.method,
  genome.ref = NULL,
  chr.list = NULL,
  exome.range = NULL
)
}
\arguments{
\item{mut.counts}{data frame of counts of mutations in each trinucleotide
context for each sample}
}
\value{
Returns the trinucleotide context fraction
}
\description{
Determines trinucleotide context fraction from a mutation counts list. Given
a data frame or .txt file or mutation counts with columns as trincucleotide
contexts in the form A[C>A]A and the rows as sample id's and a data frame or
.txt file of the times each tri nucleotide context is seen in the sequencing
region with the rows in the format ACA, the function returns a data frame of
the mutation counts normalized by trinucleotide frequency
}
\keyword{internal}
