% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_section_search.R
\name{golden.section.search}
\alias{golden.section.search}
\title{Golden section search}
\usage{
golden.section.search(f, lower, upper, tol = 1e-06)
}
\arguments{
\item{f}{function}

\item{lower}{Lower bound to search in}

\item{upper}{Upper bound to search in}

\item{tol}{How exact the answer must be}
}
\value{
Returns a value that minimizes the function input
}
\description{
Uses golden section method to search given space for value that minimizes
function given
}
\keyword{internal}
