% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSignatures}
\alias{plotSignatures}
\title{Plots the result from whichSignatures()}
\usage{
plotSignatures(sigs.output, sig.type = "SBS", sub = "")
}
\arguments{
\item{sigs.output}{The list output from whichSignatures()}

\item{sig.type}{Are SBS or DBS signatures of interest?}

\item{sub}{A character vector that specifies cancer subtype for plot title,
if wanted}
}
\value{
Plots the trinucleotide frequency for the given tumor on the top 
  panel, the calculated one on the middle panel, and the difference between 
  the two on the bottom panel.
}
\description{
Uses the output from whichSignatures() and creates a plot of the given tumor 
mutational spectrum and the calculated one
}
\examples{
plotSignatures(example.output, sub = "example", sig.type = 'SBS')
}
