% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTumor}
\alias{plotTumor}
\title{Plots a tumor profile}
\usage{
plotTumor(tumor, sub = "")
}
\arguments{
\item{tumor}{A tumor matrix}

\item{sub}{A character vector that specifies cancer subtype for plot title,
if wanted}
}
\value{
Plots the trinucleotide frequency for the given tumor
}
\description{
Uses the output from whichSignatures() and creates a plot of the given tumor 
mutational spectrum and the calculated one
}
\examples{
plotTumor(example.output[['tumor']], sub = "example")
}
