% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{randomly.generated.tumors}
\alias{randomly.generated.tumors}
\title{Values for 100 randomly generated tumor samples}
\format{A data frame of 100 rows and 96 columns}
\description{
A dataset containing the value for the 96-trinucleotide contexts of 100 
tumors generated by a random linear combination of the Nature 2013
signatures.
}
\keyword{datasets}
