% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{signatures.cosmic}
\alias{signatures.cosmic}
\title{Published Signatures from Sanger COSMIC}
\format{A data frame of 30 rows and 96 columns}
\source{
\url{http://cancer.sanger.ac.uk/cosmic/signatures}
}
\description{
A dataset containing the additional signatures identified read into R as a 
data frame. Can be used the 'signatures.ref' parameter in whichSignatures().
}
\keyword{datasets}

