% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{signatures.nature2013}
\alias{signatures.nature2013}
\title{Published Signatures from Alexandrov et al 2013}
\format{A data frame of 27 rows and 96 columns}
\source{
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl}
}
\description{
A dataset containing the published signatures from Alexandrov et al read into
R as a data frame. Can be used the 'signatures.ref' parameter in
whichSignatures().
}
\keyword{datasets}

