% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{Published Signatures from Alexandrov et al 2013}
\format{A data frame of 27 rows and 96 columns}
\description{
A dataset containing the published signatures from Alexandrov et al read into
R as a data frame. Data found:
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl}
}
\keyword{datasets}

