% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tri.counts.exome}
\alias{tri.counts.exome}
\title{The counts of every trinuclotide frequency in an exome}
\format{A data frame of 32 rows and 1 column that contains the counts}
\description{
A datset containing the number of times each trinucleotide (ex: ACA) is found
in the exome region captured by sequencing. Can be used as the
'trimer.counts.loc' parameter in whichSignatures().
}
\keyword{datasets}
