% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{updated.signatures}
\alias{updated.signatures}
\title{Published Signatures from Sanger COSMIC}
\format{A data frame of 30 rows and 96 columns}
\description{
A dataset containing the additional signatures identified
read into R as a data frame. Data found: \url{http://cancer.sanger.ac.uk/cosmic/signatures}
}
\keyword{datasets}

