% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichSignatures.R
\name{whichSignatures}
\alias{whichSignatures}
\title{Which signatures are present}
\usage{
whichSignatures(
  tumor.ref = NA,
  sample.id = NULL,
  signatures.ref = NULL,
  contexts.needed = TRUE,
  tri.counts.method = "default",
  genome.ref = NULL,
  chr.list = NULL,
  exome.range = NULL,
  associated = NULL,
  signatures.limit = NA,
  signature.cutoff = 0.06
)
}
\arguments{
\item{tumor.ref}{Either a data frame or location of input text file, where
rows are samples, columns are trinucleotide contexts}

\item{sample.id}{Name of sample -- should be the rowname of tumor.ref.
Default: `NULL`, means run deconstructSigs for all samples (rows)}

\item{signatures.ref}{Either a data frame or location of signature text file,
where rows are signatures, columns are trinucleotide contexts. Default:
NULL, means `signatures.nature2013`.}

\item{contexts.needed}{FALSE if tumor.file is a context file, TRUE if it is
only mutation counts. Default: `TRUE`.}

\item{tri.counts.method}{Set to either:
\itemize{
 \item 'default' -- no further normalization
 \item 'exome' -- normalized by number of times each trinucleotide context is
  observed in the exome \item 'genome' -- normalized by number of times each
  trinucleotide context is observed in the genome
\item 'exome2genome' -- multiplied by a ratio of that trinucleotide's
  occurence in the genome to the trinucleotide's occurence in the exome
\item 'genome2exome' -- multiplied by a ratio of that trinucleotide's
  occurence in the exome to the trinucleotide's occurence in the genome}}

\item{genome.ref}{a reference genome [BSgenome] object to define
trinucleotide's occurence.}

\item{chr.list}{the chromose the use in the analysis. Default: `paste0("chr",
c(1:22, "X", "Y"))`.}

\item{exome.range}{a [GenomicRanges] object define the exome ranges.}

\item{associated}{Vector of associated signatures. If given, will narrow the
signatures tested to only the ones listed.}

\item{signatures.limit}{Number of signatures to limit the search to}

\item{signature.cutoff}{Discard any signature contributions with a weight
less than this amount}
}
\value{
A list for every samples, the element of the list is the weights for
  each signatures, the product when those are multiplied on the signatures,
  the difference between the tumor sample and product, the tumor sample
  tricontext distribution given, and the unknown weight.
}
\description{
Determines how much of each signature is present in the sample given
}
\section{Normalization}{
 If the input data frame only contains the counts of
  the mutations observed in each context, then the data frame must be
  normalized. In these cases, the value of `contexts.needed` should be TRUE.
  \cr The parameter, `tri.counts.method`, determines any additional
  normalization performed. Any user provided data frames should match the
  format of `tri.counts.exome` and `tri.counts.genome`. \cr The method of
  normalization chosen should match how the input signatures were normalized.
  For exome data, the 'exome2genome' method is appropriate for the signatures
  included in this package. For whole genome data, use the 'default' method
  to obtain consistent results.
}

\examples{
randomly.generated.tumors <- readRDS(
    system.file("extdata", "randomly.generated.tumors.rds",
                package = "deconstructSigs")
)
test <- whichSignatures(
  tumor.ref = randomly.generated.tumors,
  sample.id = "2",
  contexts.needed = FALSE
)
}
