% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.data.R
\name{getTriContextFraction}
\alias{getTriContextFraction}
\title{Calculates trinucleotide context fraction}
\usage{
getTriContextFraction(mut.counts.ref, trimer.counts.method)
}
\arguments{
\item{mut.counts.ref}{data frame of counts of mutations in each trinucleotide
context for each sample or location where .txt file is found}

\item{trimer.counts.ref}{data frame of counts of times each trinculeotide
context is seen in sequencing area or location where the .txt file is found}
}
\value{
Returns the trinucleotide context fraction
}
\description{
Determines trinucleotide context fraction from a mutation counts list. Given
a data frame or .txt file or mutation counts with columns as trincucleotide
contexts in the form A[C>A]A and the rows as sample id's and a data frame or
.txt file of the times each tri nucleotide context is seen in the sequencing
region with the rows in the format ACA, the function returns a data frame of
the mutation counts normalized by trinucleotide frequency
}
\keyword{internal}
