% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{makePie}
\alias{makePie}
\title{Plots the weights from whichSignatures()}
\usage{
makePie(sigs.output, sub = "")
}
\arguments{
\item{sigs.output}{The list output from whichSignatures()}

\item{sub}{A character vector that specifies cancer subtype if wanted}
}
\value{
Plots a pie chart of the weights calculated in the given
tumor sample
}
\description{
Uses the output from whichSignatures() and creates a pie chart
of the weights outputted
}
\examples{
makePie(example.output)
}

