% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sample.mut.ref}
\alias{sample.mut.ref}
\title{Example input to mut.to.sigs.input()}
\format{A data frame of 657 rows and 5 column that contains example mutations 
 which could be used in mut.to.sigs.input()}
\source{
\url{http://cancergenome.nih.gov/}
}
\description{
A data frame containing example mutations that can be used as input to 
mut.to.sigs.input(). Data comes from two TCGA LUAD patients 
(http://cancergenome.nih.gov/) Contains the following columns:
}
\details{
\itemize{ \item Sample - sample name \item chr - chromosome number \item pos -
chromosome position \item ref - reference base \item alt - alternate base }
}
\keyword{datasets}
